[sdk] rearrange files so that there is one module per file
AO-6843
