+++ .yardopts
@@ -1,3 +1,4 @@
 --no-private
 --readme yardoc_frontpage.md
-lib/appoptics_apm/sdk.rb
+lib/appoptics_apm/sdk/tracing.rb
+lib/appoptics_apm/sdk/custom_metrics.rb
