+++ lib/appoptics_apm/sdk/tracing.rb
@@ -127,7 +127,7 @@ module AppOpticsAPM
       #   end
       #
       #   def handle_request_with_appoptics(request, response)
-      #     start_trace('custom_trace', nil, :TransactionName => 'handle_request') do
+      #     AppOpticsAPM::SDK.start_trace('custom_trace', nil, :TransactionName => 'handle_request') do
       #       handle_request(request, response)
       #     end
       #   end
@@ -161,7 +161,7 @@ module AppOpticsAPM
       #   end
       #
       #   def handle_request_with_appoptics(request, response)
-      #     start_trace_with_target('rails', request['X-Trace'], response) do
+      #     AppOpticsAPM::SDK.start_trace_with_target('rails', request['X-Trace'], response) do
       #       handle_request(request, response)
       #     end
       #   end
@@ -308,61 +308,5 @@ module AppOpticsAPM
 
     extend Tracing
 
-    module CustomMetrics
-
-      # Send counts
-      #
-      # Use this method to report the number of times an action occurs. The metric counts reported are summed and flushed every 60 seconds.
-      #
-      # === Arguments:
-      #
-      # * +name+          (String) Name to be used for the metric. Must be 255 or fewer characters and consist only of A-Za-z0-9.:-*
-      # * +count+         (Integer, optional, default = 1): Count of actions being reported
-      # * +with_hostname+ (Boolean, optional, default = false): Indicates if the host name should be included as a tag for the metric
-      # * +tags_kvs+      (Hash, optional): List of key/value pairs to describe the metric. The key must be <= 64 characters, the value must be <= 255 characters, allowed characters: A-Za-z0-9.:-_
-      #
-      def increment_metric(name, count = 1, with_hostname = false, tags_kvs = {})
-        with_hostname = with_hostname ? 1 : 0
-        tags, tags_count = make_tags(tags_kvs)
-        AppOpticsAPM::CustomMetrics.increment(name.to_s, count, with_hostname, nil, tags, tags_count)
-      end
-
-      # Send values for each or multiple counts
-      #
-      # Use this method to report a value for each or multiple counts. The metric values reported are aggregated and flushed every 60 seconds.
-      #
-      # === Arguments:
-      #
-      # * +name+          (String) Name to be used for the metric. Must be 255 or fewer characters and consist only of A-Za-z0-9.:-*
-      # * +value+         (Numeric) Value to be added to the current sum
-      # * +count+         (Integer, optional, default = 1): Count of actions being reported
-      # * +with_hostname+ (Boolean, optional, default = false): Indicates if the host name should be included as a tag for the metric
-      # * +tags_kvs+      (Hash, optional): List of key/value pairs to describe the metric. The key must be <= 64 characters, the value must be <= 255 characters, allowed characters: A-Za-z0-9.:-_
-      #
-      def summary_metric(name, value, count = 1, with_hostname = false, tags_kvs = {})
-        with_hostname = with_hostname ? 1 : 0
-        tags, tags_count = make_tags(tags_kvs)
-        AppOpticsAPM::CustomMetrics.summary(name.to_s, value, count, with_hostname, nil, tags, tags_count)
-      end
-
-      private
-
-      def make_tags(tags_kvs)
-        unless tags_kvs.is_a?(Hash)
-          AppOpticsAPM.logger.warn("[appoptics_apm/metrics] CustomMetrics received tags_kvs that are not a Hash (found #{tags_kvs.class}), setting tags_kvs = {}")
-          tags_kvs = {}
-        end
-        count = tags_kvs.size
-        tags = AppOpticsAPM::MetricTags.new(count)
-
-        tags_kvs.each_with_index do |(k, v), i|
-          tags.add(i, k.to_s, v.to_s)
-        end
-
-        [tags, count]
-      end
-    end
-
-    extend CustomMetrics
   end
 end
